selene: drop back to keymaster 4.0 beanpod
I hwservicemanager: Tried to start android.hardware.keymaster@4.0::IKeymasterDevice/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.

also bring the vintf entry back to 4.0

This reverts commit 8cf590e3fb4f3041a8d1fbc4d09e9c0e47a64ed1.
This reverts commit 65b7e3ff8a059dc5a23b7703b2044526adba60a9.

Signed-off-by: Hiura Mihate <tillyourteargoes@proton.me>